Add support for creating ext2 and ext3 filesystems to library
The ext* family of filesystems are distinguished from each other only by
which filesystem addon features they support -- ext3 is an ext2
filesystem with a journal, while ext4 adds more features like extents to
ext3.  It's therefore quite simple to modify the existing ext4 FS
creation code to produce other ext* family filesystems, by limiting the
features used when creating the filesystem.

Add a new make_extfs() function, which is like the existing
make_ext4fs(), except that it takes a string argument to specify the FS
type to be created: "ext2", "ext3" or "ext4".  We also need to make some
minor changes to the underlying code to account for the feature
differences between the variants.

Change-Id: Idc8bdefd93eef7c7d61dd1b70afcd3b6277f7fff